Re-added the 'punchline' onto the main page.
As talked about with jonathan this morning, I feel that we stil need an
introductory line that is interesting. We can then follow this up with a
description. A block of red writing was just not appealling at all.

I also changed the actions to be in a gray block, I'll ask van about
this later, but i think it looks a bit better than before. Ive changed
the button colors, but will get advice from van when she's here. Its
better than orange for now :)