Updated the speakers page design.
After chatting this morning, we decided that perks wasn't entierly
relevant, but we still needed something to put in that area. The
paragraph underneath wasn't really working, so i've tried putting
that content in the block instead to see how that works.

I've also made the microphone bigger and slightly simpler and moved
the text down.

there seem to have been some missing images which i've updated.