Changed navbar color and added report back
Used a color scheme designer to choose a new color for the navbar,
need to confirm with out designer if this looks ok ;)

I also made a report-back page which is useful for future attendees,
speakers and sponsors to get an idea of what last year was like.
It could probably do with some more work.