Converted speaker-info design into a template
The speaker info page was designed to be copied,
i translated it into a template and inserted template
points so that the speaker pages are literally just metadata.
In the coming commits I will be standardizing any speaker file reference
to be first_second.meta.format (eg. daniel_fiske.large.jpg for images)
so that we can reference everything using the meta data