Moved sponsor-info into template and updated sponsors
Same as was done for speakers. Moved sponsor information into yaml and updated
the design pages to be templates. Sponsors will use the same image file
path as speakers (name.meta.format). There is also a list of the
sponsors in the main site metadata for use on the sponsors and
sponsor-info pages.